Декомпозирует компонент формы отзыва
