--- src/pages/offer-page/components/feedback-form/settings.ts
@@ -1,3 +1,0 @@
-const starsCount = 5;
-
-export { starsCount };
