+++ src/pages/offer-page/components/reviews-list/reviews-list.tsx
@@ -16,4 +16,4 @@ function ReviewsList({ reviews }: ReviewsListProps): JSX.Element {
   );
 }
 
-export default ReviewsList;
+export {ReviewsList};
